requirements.txt
torch~=2.9.1+cu126
torchvision~=0.24.1+cu126
matplotlib~=3.10.7
numpy~=2.2.6
pillow~=12.0.0
opencv-python~=4.12.0.88
scikit-learn~=1.6.1
gradio~=5.49.1
tqdm~=4.67.1
seaborn~=0.13.2